Extract_data_to_csv_demo_project/dev-requirements.txt
pytest
setuptools
wheel
requests>=2.28.0
mysql-connector-python>=8.0.0
SQLAlchemy>=2.0.0
PyMySQL>=1.1.0